Implementação da junção do leveinshtein com keyboard bugada Ç_Ç

+++ src/main/java/pm/SpellCheckerLeonardo/IDistanceCalculator.java
@@ -2,5 +2,5 @@ package pm.SpellCheckerLeonardo;
 
 public interface IDistanceCalculator
 {
-	public int distance(String word1, String word2);
+	public double distance(String word1, String word2);
 }
